Authentication & authorization

+++ server/controllers/todo.js
@@ -1,10 +1,11 @@
-const { Todo } = require('../models')
+const { Todo, User } = require('../models')
 
 module.exports = class TodoController {
   static async add (req, res) {
     try {
+      let { id } = req.login
       let { title, description, status, due_date } = req.body
-      const hasil = await Todo.create({ title, description, status, due_date })
+      const hasil = await Todo.create({ title, description, status, due_date, userFK: id })
       res.status(201).json(hasil)
     } catch (err) {
       if(err.name === "SequelizeValidationError") {
@@ -15,7 +16,8 @@ module.exports = class TodoController {
   }
   static async view(req, res) {
     try {
-      let hasil = await Todo.findAll()
+      let { id } = req.login
+      let hasil = await Todo.findAll({ where: { userFK: id }})
       res.status(200).json(hasil)
     } catch (err) {
       res.status(500).json(err)
@@ -23,8 +25,9 @@ module.exports = class TodoController {
   }
   static async viewOne(req, res) {
     try {
+      let userId  = req.login.id
       let { id } = req.params
-      let hasil = await Todo.findByPk(id)
+      let hasil = await Todo.findOne({ where: { userFK: userId, id }})
       if(hasil === null) {
         throw new Error(`Row with id "${id}" not found`)
       }
@@ -35,8 +38,9 @@ module.exports = class TodoController {
   }
   static async updatePut(req, res) {
     try {
+      let userId  = req.login.id
       let { id } = req.params
-      let hasil = await Todo.update(req.body, { where: { id: id }, returning: true })
+      let hasil = await Todo.update(req.body, { where: { id, userFK: userId }, returning: true })
       if(hasil[0] === 0) {
         throw new Error(`Row with id "${id}" not found`)
       }
@@ -53,8 +57,9 @@ module.exports = class TodoController {
   }
   static async patch(req, res) {
     try {
+      let userId  = req.login.id
       let { id } = req.params
-      const hasil = await Todo.update({ status: true}, { where: { id: id }, returning: true })
+      const hasil = await Todo.update({ status: true}, { where: { id, userFK: userId }, returning: true })
       if(hasil[0] === 0) {
         throw new Error(`Row with id "${id}" not found`)
       }
@@ -71,12 +76,13 @@ module.exports = class TodoController {
   }
   static async delete(req, res) {
     try {
+      let userId  = req.login.id
       let { id } = req.params
       const deleted = await Todo.findByPk(id)
       if(deleted === null) {
         throw new Error(`Row with id "${id}" not found`)
       }
-      const hasil = await Todo.destroy({ where: { id }, returning: true })
+      const hasil = await Todo.destroy({ where: { id, userFK: userId }, returning: true })
       res.status(200).json(deleted)
     } catch(err) {
       if(err.name === "Error") {
